llm: show telemetry for repeated API calls
Currently when the LLM makes a tool call that has already been made in
the past, you'll awkwardly just see nothing - only the messages before
and after the call. This happens because telemetry is deduped
session-wide.

Now you'll see a cache hit. This is specifically only applied for LLM
tool calls, but the mechanism may have utility elsewhere.

Signed-off-by: Alex Suraci <alex@dagger.io>