chore: refactor python to remove duplications
I want to deprecate extending dagger.Enum in favor of stdlib's enum.Enum in a follow-up PR so removed str as a base class from dagger.client.base.Enum and adjusted tests to more explicitly return the value instead of the string representation. Additionallly motivated because in cattrs it bypasses the custom hook for enums, and using the one for primitive strings instead, making registration a bit more complicated.

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>