Add a module that uses /metrics with Grafana + Prometheus
We have a very simple load-test client that makes the metrics vary and
shows how everything comes together. To run it:

    dagger -m modules/metrics run up

Grafana will be available on http://localhost:3000 (default username &
password are admin:admin), and it will be already wired with
Prometheus & a Dagger Dev Engine that has `/metrics` enabled.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>