add support for prometheus metrics listener in engine (#10555)
* add support for prometheus metrics listener in engine

Starting out with just a very simple gauge for the number of connected
clients.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* support for cache size and entry count in metrics

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* limit perf impact of cache metrics by running only ever 5m

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* Add a module that uses /metrics with Grafana + Prometheus

We have a very simple load-test client that makes the metrics vary and
shows how everything comes together. To run it:

    dagger -m modules/metrics run up

Grafana will be available on http://localhost:3000 (default username &
password are admin:admin), and it will be already wired with
Prometheus & a Dagger Dev Engine that has `/metrics` enabled.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* update integ test metric name

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* appease linter

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>
Signed-off-by: Gerhard Lazu <gerhard@dagger.io>
Co-authored-by: Gerhard Lazu <gerhard@dagger.io>