ignore CA cert automatically installed by OrbStack (#10648)
* alpine: install busybox symlinks early

This handles a corner case where the engine tries to install CA
certificates after the ca-certificates package is installed but can't
because the busybox sh link hasn't been created yet.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* ignore CA cert automatically installed by OrbStack

OrbStack will by default automatically put a CA cert in containers,
including the dagger engine if the user is running on MacOS w/ OrbStack.

This results in the engine seeing a custom CA cert and deciding to
automatically install it in every container. While this works it
introduces performance overhead for a case that is unlikely to be of any
utility.

This therefore disables automatic CA cert installation when the only CA
cert found is the OrbStack one, keeping the default path case for
OrbStack users simpler and more performant.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* add changelog

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>