Various enum member followups (#10647)
* fix: avoid setting empty enum value

This could happen on old dagger versions <=v0.18.10 when the enum name
was the same as the enum value.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: patch module typedefs to correct enum default values

We need to resolve these in the typedef structure so that when we do
typedef introspection, the default values match the Names, not the
OriginalNames.

This also lets us simplify decoding the default args significantly
(which is what most of this patch updates).

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Justin Chadwell <me@jedevc.com>