chore: make DagOp helpers return non-instances
We sometimes still need to make modifications to the inputs before
dagop-ifying them, and after as well - and it's bad practice to reach
into the instance and modify things (and will soon become impossible).

We actually don't need to be accepting and returning instances in any of
these cases, so we can just remove this.

Signed-off-by: Justin Chadwell <me@jedevc.com>