chore: make DagOp helpers return non-instances (#10670)
* chore: remove unused data argument

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: remove stdin from meta mount

We can just read it from the args, since we own the call-site now.

This means we don't actually need to create a meta mount at all.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: use DagOpWrapper for various git ops

We can use the wrapper

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: make DagOp helpers return non-instances

We sometimes still need to make modifications to the inputs before
dagop-ifying them, and after as well - and it's bad practice to reach
into the instance and modify things (and will soon become impossible).

We actually don't need to be accepting and returning instances in any of
these cases, so we can just remove this.

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Justin Chadwell <me@jedevc.com>