dagql: fix case where session cache refs are never closed
If the session cache Release method was called while some calls were
still running, it was possible for them to only be appended to the list
of results after Release was already done, in which case they'd stay in
the underlying cache and leak forever.

This adds handling to track when the session cache is closed and release
any results that show up after that.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>