fix interface object loading (#10749)
* fix interface object loading

There was a very subtle issue where `ObjectResult`s were accidentally
being "downgraded" to `Result` when dagql call internals were attempting
to unwrap them. This was due to unintentionally calling DerefValue on
the embedded `Result` of `ObjectResult` and returning the received,
which was a `Result` rather than full `ObjectResult`.

This broke interface loading specifically because we currently rely on
some objects already being selectable rather than needing to load their
type information; if the object is an interface value from some other
module unknown to the current one we can't load the type and error out.

This technically restores us back to the previous way this all worked,
but does still feel like very thin ice. Need this regression fixed
before the release though, so kicking the can on reforming the whole
interface system here.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* add integ test coverage

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>