exp: allow to run dagger 'tasks' from a dagger.yaml file
This is experimental, just to see if that could be a good idea or not.

All start with a `dagger.yaml` file you can put in your project.
The `dagger.yaml` file is composed of two sections: `vars` and `tasks`.

`vars` allow to define some variable (could be basic variable like
strings, or more complex using any dagger primitives).

For instance this is declaring a Go version and defining a `ctr`
variable representing a basic container to work on a go project:

    vars:
      go_version: 1.24.4
      ctr: >-
        $(container |
        from golang:$go_version-alpine |
        with-exec go install golang.org/x/tools/cmd/goimports@latest |
        with-exec go install mvdan.cc/gofumpt@latest |
        with-workdir /go/src |
        with-mounted-file /go/src/go.mod go.mod |
        with-mounted-file /go/src/go.sum go.sum |
        with-mounted-cache /root/.cache $(cache-volume go-cache) |
        with-mounted-cache /go/pkg/mod $(cache-volume go-pkg-mod) |
        with-exec go mod download |
        with-mounted-directory /go/src .)

Those variables will be translated as dagger shell variables, so
available as `$go_version` and `$ctr`.

`tasks` allow to define some tasks to run. They have a name and a body
that is a dagger shell command. This command can use previously defined
variables.

For instance:

    tasks:
      terminal: >
        $ctr | terminal

      go:test: >
        $ctr |
        with-exec -- go test -coverpkg=./... -coverprofile=.coverage -shuffle=on ./... |
        stdout

This will define two tasks, one `terminal` that will open a terminal on
the dev environment defined by `$ctr` and the second will run tests
inside this container.

Usage:

- dagger task: list all available tasks
- dagger task TASK: will run the corresponding task by concatenating all
  the vars and the task body and run it as a dagger shell command

To test it, you can use `daggerfile` branch of `eunomie/docker-runx`
project that contains a quick example: https://github.com/eunomie/docker-runx/tree/daggerfile

Clone the project, switch to daggerfile branch and run `dagger task`
then `dagger task go:test`

Signed-off-by: Yves Brissaud <yves@dagger.io>