just log session closed errors for now rather than return them
Despite squashing a number of occurences, we are still seeing occasional
flakes related to using stale sessions. These errors do mean something
unexpected is happening with the lifecycle of session caches and should
be an error in the long term, but in the short term it's not strictly
necessary to return them; things *should* generally work.

This update thus downgrades returning the error to just logging it,
while still avoiding memory leaks previously caused by this problem.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>