experiment: static (...dynamic?) tools
Rather than call_method and friends, expose tools directly as LLM tools,
just as before.

We desperately need a name for this though: it's static from the POV
that you don't dynamically select and invoke methods, but it's dynamic
from the POV that the available tools change as you discover more object
types. (Which wasn't exactly a goal - I originally wanted to constrain
it to the inputs - but need to explore that more.)

Signed-off-by: Alex Suraci <alex@dagger.io>