fix loading of optional secrets from module objects
When a module object had a Nullable field for an object type, we were
subtly setting the ID for the field to the ID that selects the field,
rather than the ID for the object it wraps (when the nullable is set).

This didn't impact too much other than secrets, where it could break
loading in certain situations when the secret would be attempted to be
loaded *from the field selector ID* rather than the underlying ID for
the secret from the core API.

This fixes the issue by ensuring Nullables wrap the underlying Result,
which allows us to leave the "real" ID intact and used instead of the ID
that just selects the nullable field.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>