fix race condition in cache when context canceled
CI hit an error in the cache unit tests, locally I could repro it in ~2
runs when running w/ `-count=100000`, so unclear if it was really
causing any genuine problems.

Was a legit problem though; with the right timing it was possible to
remove a call from internal maps when there were waiters remaining,
which would result in extra calls appearing.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>