Don't convert module name to main object name
We're setting the main object automatically at import time through the decorator, but keep the validation during only registration because it always needs to happen first, before invocation.

Besides, we only really need to know what the main object is during registration in order to register the constructor, although we could register constructors for all types since the engine will discard any that are not for the main object.

So apart from a slight performance benefit we still validate the main object to give a specific error message about decorating the class with `@dagger.object_type` which if left to the engine the error message will be more general and thus slightly less useful.

Finally, the runtime module already does this and it's best to make sure they're the same.

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>