fix: go client generator now works on empty dir (#10821)
Previously, generating a go client in an empty directory led to an error:

```
mkdir test && cd test
dagger init
dagger client install go
```

Because the go client generator assumed that the module was aside a
go module, same as module generation before we generate a default
go.mod.

This PR fixes that issue by: generating a default go.mod based on the
module name if no go.mod exist in the current directory.
This allows an user to create a submodule in a go project and import
the client from there.

Also fixes an issue where the client installation failed if no go file
were found in the module directory.

Add integration tests to verify the new behaviour.

Signed-off-by: Tom Chauveau <tom@dagger.io>