fix: allow to dagger init a go module with self types
In case of an empty module during the typedefs phase, generate
the default template so that types from this template are returned
and will be handled by codegen. That way, with a `dagger init --sdk=go`
the types of this "empty" module are available from the start and don't
require a `dagger develop` to be available.

This also adds the introspection schema json file to the typedefs
function of the generator. This allows to generate enough code
so that the module initialisation can correctly load the packages
and types and generates the right types.

Signed-off-by: Yves Brissaud <yves@dagger.io>