fix: add --first-parent for counting rev-list
We always should have been using `--first-parent` here, without that, we
always get *all* the parents, which counts all parents of a merge
commit, which isn't really useful for estimating the depth.

This is also really just a guess - it's possible that different parents
can end up with different depths. This means it's *theoretically*
possible to get the wrong results here.

Signed-off-by: Justin Chadwell <me@jedevc.com>