python: improve error reporting in modules (#10532)
* python: improve error handling and reporting

- Logs exported to OTel via OTLP; should be converted to stderr output by Dagger CLI.
- Logging exceptions closest to where they happen so the the stack trace is shown first, with less noise in the end.
- Full stack trace added as an extension to `dagger.Error`, using same key names from OTel convention.
- Exceptions re-raised as custom errors because:
    - Bubbles up the exit condition
    - User facing error message, hoisted in telemetry UI
    - Lets the general error handler know it’s already been logged to reduce noise

* Fix unhandled exception in task group
* Capture exception on building the query builder's DSL
* Don't convert module name to main object name

We're setting the main object automatically at import time through the decorator, but keep the validation during only registration because it always needs to happen first, before invocation.

Besides, we only really need to know what the main object is during registration in order to register the constructor, although we could register constructors for all types since the engine will discard any that are not for the main object.

So apart from a slight performance benefit we still validate the main object to give a specific error message about decorating the class with `@dagger.object_type` which if left to the engine the error message will be more general and thus slightly less useful.

Finally, the runtime module already does this and it's best to make sure they're the same.

* Udpate golden test
* Show traceback on json decoding
* Add change log
* Reduce noise in API errors
* Fix tests
* Change to breaking
* Add a few tests

---------

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>