feat: support OIDC auth with the Dagger Cloud API
The problem: public repositories today must expose their credentials if
they wish to send telemetry or do any kind of operation against our
cloud API. We are currently exposing our own DAGGER_CLOUD_TOKEN but
since its only telemetry we don't worry much about it. Now that we are
slowly introducing compute features, its critical that we have a
mechanism to not expose credentials that could be used to provision
compute in our behalf.

This implements an OIDC authentication flow against our cloud API with
GitHub as the first provider.

Signed-off-by: Matias Pan <matias@dagger.io>