chore: defer session cleanups while in no-exit state
By using the `-E` flag, or the `E` hotkey, we can enter a special TUI
no-exit state.

While in this state, even after the dagger operation is completed, we
keep the TUI open and running. However, previously, we were completely
shutting down the session and the telemetry.

For some interesting future operations, we want to defer these callbacks
until we're quitting the TUI completely - we need an active client.

Signed-off-by: Justin Chadwell <me@jedevc.com>