re-add --restart=always to docker run for engine
This option got dropped during driver refactoring, which seems to have
caused engine start in CI to fail >50% of the time due to `sed: write
error` when the engine tries to setup cgroups during its entrypoint.
* https://github.com/dagger/dagger/issues/7785

Previously that would happen, the engine would restart and then succeed
since the sed error is ephemeral. However, without the option the engine
just never starts.

This change just re-adds that to the docker run command line.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>