feat: add ability to combine outputs in WithExec (#10924)
A new function `combinedOutput` is available, in addition to `stdout`
and `stderr`. It means you can pick the one you want depending on the context.
For instance to pick the combined output if the command failed.

    $ dagger -q <<EOF
    container | from alpine |
    with-new-file /test.sh "echo out; echo err >&2" |
    with-exec -- sh /test.sh |
    combined-output
    EOF

    out
    err

Warning: as always while working with streams, order in the output is not guaranteed.

fixes #3496

Signed-off-by: Yves Brissaud <yves@dagger.io>