dagql: optimize InputSpecs.Inputs
Before this, a pprof cpu profile of the engine while running
TestTelemetry showed that 7-10% of our cpu time was spent in
dagql.InputSpecs.Inputs. Most of this was due to the `seen` map creation
and read/writes.

Given the small size of `n` for the function, using an O(n^2) scan of
the slice to check for whether an arg was already seen is actually much
more performant than the map.

After that fix, Inputs went down to only consuming 0.1% of the cpu time.
It could probably be optimized further by e.g. caching results but given
its no longer a bottleneck that extra complication doesn't seem worth it
at the moment.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>