optimize ID calcDigest
After the previous optimizations in this commit series, ID.calcDigest
was using ~13% of cpu time during runs of TestTelemetry.

The biggest bottleneck in it was (surprisingly) that it grew the stack
enough to trigger a resize, which the go runtime spent a lot of work
doing. Since `calcDigest` is called *a lot*, that resizing added up
quite a bit and was much more expensive than the actual marshaling and
hashing.

To fix the usage of stack size:
1. The xxh3 hash object is stored in a sync.Pool. That hasher includes a
   >1kb array (not slice) and it appears the call to xxh3.New() was
   always being inlined, so it didn't end up in the heap. Using a
   sync.Pool ensures it is heap-stored while also saving work
   allocating/GC'ing the object.
2. After the above, the protobuf marshal code started being the one
   triggering stack size growth. To fix that, we now roll our own by
   just writing each relevant pb field to the hash. This requires some
   tedious manual effort + maintaince, but given the perf benefits seems
   worth it.

After all the above, `calcDigest` went from 13% of cpu usage down to
0.9%.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>