Grab bag of engine optimizations (#10965)
* dagql: optimize InputSpecs.Inputs

Before this, a pprof cpu profile of the engine while running
TestTelemetry showed that 7-10% of our cpu time was spent in
dagql.InputSpecs.Inputs. Most of this was due to the `seen` map creation
and read/writes.

Given the small size of `n` for the function, using an O(n^2) scan of
the slice to check for whether an arg was already seen is actually much
more performant than the map.

After that fix, Inputs went down to only consuming 0.1% of the cpu time.
It could probably be optimized further by e.g. caching results but given
its no longer a bottleneck that extra complication doesn't seem worth it
at the moment.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* optimize ID calcDigest

After the previous optimizations in this commit series, ID.calcDigest
was using ~13% of cpu time during runs of TestTelemetry.

The biggest bottleneck in it was (surprisingly) that it grew the stack
enough to trigger a resize, which the go runtime spent a lot of work
doing. Since `calcDigest` is called *a lot*, that resizing added up
quite a bit and was much more expensive than the actual marshaling and
hashing.

To fix the usage of stack size:
1. The xxh3 hash object is stored in a sync.Pool. That hasher includes a
   >1kb array (not slice) and it appears the call to xxh3.New() was
   always being inlined, so it didn't end up in the heap. Using a
   sync.Pool ensures it is heap-stored while also saving work
   allocating/GC'ing the object.
2. After the above, the protobuf marshal code started being the one
   triggering stack size growth. To fix that, we now roll our own by
   just writing each relevant pb field to the hash. This requires some
   tedious manual effort + maintaince, but given the perf benefits seems
   worth it.

After all the above, `calcDigest` went from 13% of cpu usage down to
0.9%.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* fix local source stat hash byte append

We were appending to a []byte but not using the newly updated slice, so
the data written was just being overwritten each time. Surprisingly,
this didn't *noticeably* break anything yet? I just noticed it while
working on something else that also used the same Append methods.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* honor docker credentials during TestTelemetry to avoid rate limits

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>