fix: ensure consistency when excluding gitignore
This was a weird one - if you forcefully excluded loading gitignore
while attempting to apply the gitignore, you could get inconsistent
results.

Assuming strictly serial uploads this was fine, but in parallel, we're
very careful to ensure that the shared file store is in a consistent
state.

For example, suppose:
- From /src, we upload a large a.txt (using include=a.txt)
- From /src in parallel, we upload a .gitignore containing a.txt (using
  exclude=.gitignore, and applying gitignore)
    - The expected outcome is to not upload a.txt
- But, because a.txt is in the local copy and the gitignore is *not* (we
  skipped uploading it), then we incorrectly will copy a.txt to our
  snapshot.

The *core* problem here is that we are effectively applying
includes+excludes and gitignores *twice*. First when uploading, and
second when we copy to our snapshot.

We can fix this! Instead of applying the patterns again on the second
copy, we know *exactly* what files need to be copied (based on the
transfer results), so can just include those.

Signed-off-by: Justin Chadwell <me@jedevc.com>