fix: additional local/remote consistency
Update syncParentDirs to always sync over gitignores. This ensures that
the gitignore patterns between local/remote *always* match.

This wasn't previously true, and so could have potentially meant that we
were removing files that were still being used. Imagine the scenario:
- We upload a large a.txt, excluded by gitignore, but gitignore not applied
- We upload something else, and apply gitignore (but avoid transferring
  the gitignore, by excluding the gitignore file in a pattern)
    - This will remove a.txt, because now local has no idea that it
      should be excluded.
- If we upload the same a.txt again, we'll have to do it again from
  scratch.

This is all weird, so instead of this, we'll just first sync *all* the
gitignores, so that the context perfectly matches. This actually
resolves the same problem as the prior commit - but also resolves the
above issue. Technically, this would work by itself, but the additional
copy is still confusing and should still be removed.

Signed-off-by: Justin Chadwell <me@jedevc.com>