ref: split moduleTypeDefs and self calls
In case `moduleTypeDefs` exists in the SDK, use it to get the module's
types. But if the self calls experimental flags, simply replace the
previous moduleRuntime call, do not use the results as its own dependency.

This allows to remove the previous duplication between TypeDefCode and TypeDef
in go codegen for instance, and to benefit from the moduleTypeDefs work
without the self calls that needs more work to be ready for all cases.

Signed-off-by: Yves Brissaud <yves@dagger.io>