chore: swap order of filtering for performance
It turns out to be much *cheaper* to apply gitignores after
include/exclude.

e.g. we might come to exclude a huge directory - in that case, it's not
worth looking for gitignore patterns in there. So we can filter that out
before.

But we still use the original filesystem for matching! That way, we
can't accidentally exclude the gitignore.

Signed-off-by: Justin Chadwell <me@jedevc.com>