fix: preserve call ids during id extraction
We should preserve the call digests as we encounter them in the call
we're attempting to marshal - this is important because the output
digest of a call can be different from the input digest!

For an example:

    $ dagger core -E container from --address=alpine with-mounted-directory --path=/src --source=. with-exec --args ls,-lh,src stdout

It's not possible to debug the `withMountedDirectory` (or anything after
it) without this patch applied.

That's because the `withMountedDirectory` has a different output digest.
The logic in `db.Call` handles this, but we need to make sure to keep
the ids consistent, or we produce a garbage ID with dangling references.

Signed-off-by: Justin Chadwell <me@jedevc.com>