optimize out some complex introspection queries (#10967)
Introspection queries go through dagql and trigger a pretty insane
amount of calls to dagql.Server.Resolve each time a client connects
or a new server is made.

This isn't a currently bottleneck on machines with enough CPU, but it
makes pprof traces almost unusable at times due to the sheer number of
goroutines. I would not be shocked if there are scenarios where it is a
genuine bottleneck too (especially in cases of lots of concurrent
clients).

This change updates some of our internal calls to introspection queries
to be implemented directly rather than as an actual dagql query. It also
tweaks dagql.Server.Resolve a bit to only use separate goroutines when
there are multiple selections, otherwise just calling directly.

To test the difference, the number of times Resolve gets called when
running `dagger -m ./modules/alpine call --help`:
* Before: 23712
* After: 3548

There are also pprof traces that previously were not even loadable in my
browser that I can now open successfully.

This isn't a complete fix; the CLI typedef queries appear to be another
significant source that could be optimized out. And like I said there
isn't a immediately obvious bottleneck getting cleared here. But still
seems worth it overall.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>