llm: durable to LLMs invoking non-namespaced tools
gemini-2.0-flash struggles because Claude tool descriptions mention
naked tool names like "Read" - but we still want to namespace tools. to
handle this, we'll just check both the exact match and un-namespaced
match. if there's only one match, we'll use it; if there are multiple,
we'll give a clear, helpful error.

Signed-off-by: Alex Suraci <alex@dagger.io>