dedupe concurrent dagql calls within same client (#11070)
Previously on this journey, we disabled deduping of concurrent dagql
calls and the related "edge-merging" feature of buildkit. Motivation
being that we did not have any handling of a number of corner cases like
a client canceling their request that got deduped with other clients.

We are still leaving buildkit edge merging off forever, but this change
reintroduces de-duping to the dagql cache with the new twist that it
only occurs within calls that belong to the same client.

Now if a single client has two concurrent calls that are the same and
still running, they will be deduped. If two separate clients are making
the same concurrent call, they will still run separately (and one will
"win" in terms of what's written to cache).

The isolation to a single client avoids the previous problems with
client cancelation.

The benefit is more predictable and improved performance. In particular,
I noticed while working on something else that the way we load module
deps in parallel and recursively was resulting in a lot of overlapping
concurrent calls in the same client (using the dagger repo as an
example, which has a complicated module dependency dag).

The change here gets rid of that duplication. This practically speaking
only appears to shave ~2s off module loading time for me. However, it
makes it much easier to reason about how often various pieces of that
code will run (at most once, per-client), which in turn makes the whole
thing easier to think about and less delicate in terms of future changes
impacting performance.

The module loading example is just one, I'm sure there's other
real-world use cases that result in concurrent overlapping dags being
solved by a single client; any code that does solves in parallel has a
decent chance of benefitting.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>