chore: allow pinning git refs
This allows refs to be arbitrarily pinned to any commit - this is an
internal API designed to be used for module pinning.

Essentially, all we do is heavily refactor the internals so that we're
not reliant on the Ref pointing at the SHA. We already reconfigure all
the relevant pointers, so we do that here too.

Signed-off-by: Justin Chadwell <me@jedevc.com>