ts(introspector): drop empty deprecation tags when scanning decorators
TypeScript's JSDoc previous logic was treating no @deprecated tag and empty ones the same way: by assigning an empty string to the deprecated field.

Now that we've made deprecated nullable across the ts sdk, a tag empty constructs an empty string and if some text was added, it is appended too.

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>