llm: don't namespace mcp tools
This just isn't how the world works at the moment, and causes some
models (Gemini) to just not work. The problem is LLM tool descriptions
sometimes reference adjacent tools, which convinces the model to call
them by the un-namespaced name, which causes Gemini to hit a malformed
tool call.

Instead, we'll just do last-write-wins, and adopt an OrderedSet model
for keeping track of the toolset.

Also, switched method tools from TypeName.foo to TypeName_foo so we
don't need to do clunky re-naming.

Signed-off-by: Alex Suraci <alex@dagger.io>