fix: avoid recomputing content hashes for git tree hashes
When creating a clone of a private repo we do a fun little dance to
mixin the content hash of the clone so that we avoid mixing up two
repos with the same content, but different URLs. This scenario appears
frequently with out module tests (with mirrored versions of the same
repo in multiple places).

When we introduced `maintainContentHashing`, the intention was to ensure
that a `Directory` that was previously content-hashed (e.g. on load from
a client using `Host.directory`) *stayed* content content-hashed when we
select one of it's subdirectories. This is a relatively cheap operation,
because we should have already created a cache context.

However - these two ideas conflict. To see this:

    $ dagger shell

    ▶ git "git@github.com:dagger/dagger.io" | head | tree
    Directory@xxh3:85f08c1a0f41f188

    ▶ git "git@github.com:dagger/dagger.io.git" | head | tree
    Directory@xxh3:6cf30aa8e3e2274e

    ▶ git "git@github.com:dagger/dagger.io" | head | tree | directory api
    Directory@sha256:f3e0e7d9fec74300ec961a442ad139b0d3b1b7262b3c217a47af4d1a8b472c97

    ▶ git "git@github.com:dagger/dagger.io.git" | head | tree | directory api
    Directory@sha256:f3e0e7d9fec74300ec961a442ad139b0d3b1b7262b3c217a47af4d1a8b472c97

Note how the top-level directory uses a custom dagql hash, while the
selection uses a content hash; now these two *different* URLs which
used to result in two different IDs now result in a collision. This
causes issues downstream (not sure why, struggling to reproduce - I'm
not sure why this would end up being computed again, since this collision
should only be reachable after we've actually cloned the repo, however,
this could happen if the buildkit cache is tidied at the right point
*during* a test run).

The fix here is to narrow our definition of a content hash - the
buildkit `contenthash` package always creates SHA256 hashes, so we
should *only* maintain content hashing if we know that a SHA256 hash was
set (if it's creating using dagql.From, it'll be an XXH3 hash, and we've
deliberately overwritten it).

This feels hacky - but I don't have a super clever suggestion here, it
feels like the correct solution is to work out *how* to allow these
content hash collisions (so it's not an issue in the first place), but
as of the current state of dagql<->buildkit cache, I'm not sure how
feasible that is.

Signed-off-by: Justin Chadwell <me@jedevc.com>