tests: remove standalone services tests
These are pretty old - also, from my memory they haven't caught any
issues. Our integration tests are really the authority here.

However, they have been needed to be modified to keep up to date, and
are the reason that the `Startable` interface exists. If we remove
`fakeStartable`, then we don't need that, which ends up simplifying a
lot of code.

Signed-off-by: Justin Chadwell <me@jedevc.com>