ref: use a container for ModuleDefs implementation
This will allow to control how the WithExec call is made, especially
this will allow to inject the module information during the call.
But this needs to happen from the engine, not from the SDKs.

To inject module data will allow to set the right engine version
at a moment the module is still not yet loaded.

Signed-off-by: Yves Brissaud <yves@dagger.io>