fix: python moduleDefs implementation
Always codegen. This is the only way, with the actual implementation,
to have access to the SDK for instance.

The test is wrapping the Python SDK, in a module written in Python.
The code is calling dag.PythonSDK. But this module exists only because it
has been loaded, it's not part of the default core types.
This shouldn't be an issue, except the way the Python SDK works
is moduleDefs is using moduleRuntime container.

The ultimate solution is to cut the link between moduleDefs and moduleRuntime,
to have better performances, but in the meantime this can be a solution.

Signed-off-by: Yves Brissaud <yves@dagger.io>