fix handling of private images in terminal
The terminal implementation was not providing a session group when
creating new mounts. This meant that if an input mount was derived from
a still-lazy private image, no auth was available and pulling the image
would fail.

This was exacerbated by a missing check for `err != nil`, which meant we
went ahead and tried to start the container even though all the mounts
were nil, which crashed the engine.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>