llm: propagate env workspace through Query
Previously, the LLM/MCP implementation manually filled in arguments that
had a `@defaultPath` annotation. This worked to some extent, but did
_not_ work if the module function then called another module that used
contextual args - those would still just load from `Host`!

In other words, you'd end up with a call stack like this:

    LLM
    Dev(Env.workspace).test                      # WORKS
    DaggerDev(Host.directory(...)).test          # BROKEN (stale data)
    DaggerEngine(Host.directory(...)).container  # BROKEN (stale data)

Now we propagate the `Env` into `Query`, where it gets picked up
directly by the `defaultPath` handling, which uses it instead of `Host`
if it's present.

With that change, the `Env` workspace propagates cleanly throughout the
call stack:

    LLM
    Dev(Env.workspace).test                            # WORKS
    DaggerDev(Env.workspace.filter(...)).test          # WORKS
    DaggerEngine(Env.workspace.filter(...)).container  # WORKS

Signed-off-by: Alex Suraci <alex@dagger.io>