feat: add withError
Add withError to Container and Directory objects,
to allow to raise errors in a self chainable way.

If the error message is empty, the operation will be ignored.

If the error message is not empty, the error will be raised
when the call to withError will be resolved.

This allows to keep chaining calls, and especially to use it
with 'with' call.

    ctr = ctr.
    	With(func(r *dagger.Container) *dagger.Container {
    		ctr, ok = doSomething(r)
    		if !ok {
    			return r.WithError("not ok")
    		}
    		return ctr
    	})

Signed-off-by: Yves Brissaud <yves@dagger.io>