feat: add withErrorMessage
Add withErrorMessage to Container and Directory objects,
to allow to raise errors in a self chainable way.

If the error message is empty, the operation will be ignored.

If the error message is not empty, the error will be surfaced
and subsequent calls will be canceled.

This allows to keep chaining calls, and especially to use it
with 'with' call.

    ctr = ctr.
    	With(func(r *dagger.Container) *dagger.Container {
    		ctr, ok = doSomething(r)
    		if !ok {
    			return r.WithErrorMessage("not ok")
    		}
    		return ctr
    	})

Signed-off-by: Yves Brissaud <yves@dagger.io>