feat: expose introspection-schema-json file to module-source
This will help to develop SDKs, but this will also allow to invert
the way SDKs are working with this introspection json file as the SDK
will now be able to ask for the introspection file, when needed, and not
upfront. The goal is to be able to limit cache invalidation by not
requesting data that are not needed at some point.

Signed-off-by: Yves Brissaud <yves@dagger.io>