fix: apple container engine start (#11243)
When we provision engine, we try to always 'start' existing container
before to check the state of the container.
This works well with Docker as a 'docker start' on an already existing
and started container has no impact.

But with apple containers, to 'container start' an already existing
container will stop and remove the container.

Instead, check the running status of the container, and if the container
is already running, to not attempt to start it.

That way, from one call to the other, the container is kept running.

Fixes: #11242

Signed-off-by: Yves Brissaud <yves@dagger.io>