fix(typescript-sdk): use `npm install` instead of `npm ci`
Fixes an issue raised by @tibor about module execution.
`npm ci` would lead to a failure if the package.lock is out of date.
That can happen for old modules that haven't been updated.

This can also happen for other package managers so we also disable it.

Benchmark shows a performance drawback of 3% with that changes.
It's acceptable

Signed-off-by: Tom Chauveau <tom@dagger.io>