fix: ensure that __schemaJSONFile returns the deprecationReason
it's taken me some time to understand why the templates vs the
introspection query were seeing a different schema.

It turns out that we filter out information to SDKs on this query and
that we were not passing all the fields. Here, we properly transfer the
dagql fields to the returned go types

I also added the json debug helper to easily understand and debug what
the template sees at a given point in time

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>