refactor: move CacheSpec to FieldSpec
Previously this only existed on Field[T] which meant we can't access it
generically - we have to access the field on the concrete Class[T],
which we don't always know.

With this change we can just grab the cache config off the FieldSpec and
use it to e.g. fill out updated contextual arguments.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>