refactor: break up ID.Append
This was already getting a bit unwieldy as we've been gradually adding
bits and bobs to the protocol, and it was also used to represent all
forms of modification.

Refactored into an `IDOpts` pattern, and added `ID.With(...IDOpts)` for
applying selective modifications to an ID, rather than having to
re-`Append` from its receiver.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>